+++ dart/dynamics/Skeleton.cpp
@@ -2267,6 +2267,9 @@ void Skeleton::registerBodyNode(BodyNode* _newBodyNode)
 
   _newBodyNode->incrementVersion();
   _newBodyNode->mStructuralChangeSignal.raise(_newBodyNode);
+  // We don't need to explicitly increment the version of this Skeleton here
+  // because the BodyNode will increment the version of its dependent, which is
+  // this Skeleton.
 }
 
 //==============================================================================
@@ -2438,6 +2441,12 @@ void Skeleton::unregisterBodyNode(BodyNode* _oldBodyNode)
   }
 
   updateTotalMass();
+
+  _oldBodyNode->incrementVersion();
+  _oldBodyNode->mStructuralChangeSignal.raise(_oldBodyNode);
+  // We don't need to explicitly increment the version of this Skeleton here
+  // because the BodyNode will increment the version of its dependent, which is
+  // this Skeleton.
 }
 
 //==============================================================================